# Adding batch processing start/end for install/uninstall packages to improve performance