Adding batch processing start/end for install/uninstall packages for ms build projects which improved performance by ~25%

Fix https://github.com/NuGet/Home/issues/2903

Below numbers are taken while updating 35 packages across 6 different types of projects in a solution:
Without batch  With batch processing
92 seconds	   66 seconds
87 seconds	   67 seconds
87 seconds	   58 seconds
91 seconds	   60 seconds
89 seconds	   64 seconds