@jainaashish 2016-06-21
Since there is no 3.4.5 now so closing this PR